# services/@tables/service/spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tables Service API",
    "description": "Dynamic data tables API with Yjs CRDT real-time collaboration",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/tables/{tableId}/columns": {
      "post": {
        "operationId": "addColumn",
        "tags": [
          "columns"
        ],
        "description": "Add a new column to a table",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "dataType"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Column name",
                    "minLength": 1
                  },
                  "dataType": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "text"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "number"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "boolean"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "date"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "select"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tableId",
            "required": true,
            "description": "Table ID"
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "dataType",
                    "position"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Column ID (UUID)"
                    },
                    "name": {
                      "type": "string",
                      "description": "Column display name"
                    },
                    "dataType": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "text"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "number"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "boolean"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "date"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "select"
                          ]
                        }
                      ]
                    },
                    "position": {
                      "type": "number",
                      "description": "Column position (0-indexed)"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tables/{tableId}/columns/{columnId}": {
      "patch": {
        "operationId": "updateColumn",
        "tags": [
          "columns"
        ],
        "description": "Update a column (rename or change type)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "New column name"
                  },
                  "dataType": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "text"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "number"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "boolean"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "date"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "select"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tableId",
            "required": true,
            "description": "Table ID"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "columnId",
            "required": true,
            "description": "Column ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "dataType",
                    "position"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Column ID (UUID)"
                    },
                    "name": {
                      "type": "string",
                      "description": "Column display name"
                    },
                    "dataType": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "text"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "number"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "boolean"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "date"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "select"
                          ]
                        }
                      ]
                    },
                    "position": {
                      "type": "number",
                      "description": "Column position (0-indexed)"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteColumn",
        "tags": [
          "columns"
        ],
        "description": "Delete a column",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tableId",
            "required": true,
            "description": "Table ID"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "columnId",
            "required": true,
            "description": "Column ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tables/{tableId}/columns/reorder": {
      "post": {
        "operationId": "reorderColumns",
        "tags": [
          "columns"
        ],
        "description": "Reorder columns in a table",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "columnIds"
                ],
                "properties": {
                  "columnIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Column IDs in new order"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tableId",
            "required": true,
            "description": "Table ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "name",
                      "dataType",
                      "position"
                    ],
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "Column ID (UUID)"
                      },
                      "name": {
                        "type": "string",
                        "description": "Column display name"
                      },
                      "dataType": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "text"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "number"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "boolean"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "date"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "select"
                            ]
                          }
                        ]
                      },
                      "position": {
                        "type": "number",
                        "description": "Column position (0-indexed)"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3005",
      "description": "Development server"
    }
  ]
}